# apps/fastify/api/menu/swagger.yaml
openapi: 3.0.3
info:
  title: Menu API
  description: 菜单管理 API 文档
  version: 1.0.0
components:
  schemas: {}
paths:
  /api/menus:
    get:
      parameters:
        - schema:
            type: string
          in: query
          name: parentId
          required: false
        - schema:
            type: string
            enum:
              - 'true'
              - 'false'
          in: query
          name: flat
          required: false
      responses:
        '200':
          description: Default Response
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - label
              properties:
                name:
                  type: string
                label:
                  type: string
                path:
                  type: string
                icon:
                  type: string
                order:
                  type: integer
                parentId:
                  type: string
        required: true
      responses: {}
  /api/menus/root:
    get:
      responses:
        '200':
          description: Default Response
  /api/menus/{id}:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses: {}
    put:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                label:
                  type: string
                path:
                  type: string
                icon:
                  type: string
                order:
                  type: integer
                parentId:
                  type: string
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses: {}
    delete:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses: {}
  /api/menus/{id}/children:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses: {}
servers:
  - url: http://localhost:3000